version 0.2 release

+++ prog.sh
@@ -3,12 +3,18 @@
 # bash script to run Robinhood updates
 # 1 - update update.txt
 update_file=/home/tyler/robinhood/update.txt
-rm $update_file
-# touch $update_file
-# chmod 666 $update_file
+if [ -f $update_file ]
+   then 	
+      rm $update_file
+      touch $update_file
+      chmod 666 $update_file
+fi
 # 2 - run my_stocks.py
-python3 my_stocks.py > update.txt
+python3 /home/tyler/robinhood/my_stocks.py > update.txt
 # 3 - run mail.py 
-python3 mail.py
+# python3 /home/tyler/robinhood/mail.py
+
+# return update.txt content
+cat update.txt
